+++ firebase-annotations/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
 
 - [changed] Hid Executors from public API, as they are intended to be internal anyhow.
+- [feature] InternalForInheritanceFirebaseApi annotation added.
+  ([#8603](https://github.com/firebase/firebase-android-sdk/pull/8603))